requirements.txt
beautifulsoup4==4.14.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
contourpy==1.3.3
coverage==7.13.4
curl_cffi==0.13.0
cycler==0.12.1
et_xmlfile==2.0.0
execnet==2.1.2
fonttools==4.62.0
frozendict==2.4.7
idna==3.11
iniconfig==2.3.0
kiwisolver==1.5.0
llvmlite==0.44.0
matplotlib==3.10.8
multitasking==0.0.12
numba==0.61.0
numpy==2.1.3
openpyxl==3.1.5
packaging==26.0
pandas==3.0.1
peewee==4.0.1
pillow==12.1.1
platformdirs==4.9.4
pluggy==1.6.0
protobuf==7.34.0
pyarrow==23.0.1
pycparser==3.0
Pygments==2.19.2
pyparsing==3.3.2
pytest==9.0.2
pytest-cov==7.0.0
pytest-xdist==3.8.0
python-dateutil==2.9.0.post0
pytz==2026.1.post1
PyYAML==6.0.3
requests==2.32.5
scipy==1.17.1
six==1.17.0
soupsieve==2.8.3
typing_extensions==4.15.0
urllib3==2.6.3
websockets==16.0
yfinance==1.2.0
